Polish documentation
