+++ core/engine/src/builtins/array/mod.rs
@@ -41,6 +41,9 @@ mod array_iterator;
 use crate::value::JsVariant;
 pub(crate) use array_iterator::ArrayIterator;
 
+#[cfg(feature = "experimental")]
+mod from_async;
+
 #[cfg(test)]
 mod tests;
 
@@ -663,440 +666,6 @@ impl Array {
 
         // 2. Return ? IteratorClose(iteratorRecord, error).
         iterator_record.close(error, context)
-    }
-
-    /// [`Array.fromAsync ( asyncItems [ , mapfn [ , thisArg ] ] )`][spec]
-    ///
-    /// The `Array.fromAsync()` static method creates a new,
-    /// shallow-copied Array instance from a list or iterator of Promise-like values.
-    ///
-    /// More information:
-    ///  - [ECMAScript reference][spec]
-    ///
-    /// [spec]: https://tc39.es/proposal-array-from-async/#sec-array.fromAsync
-    #[cfg(feature = "experimental")]
-    pub(crate) fn from_async(
-        this: &JsValue,
-        args: &[JsValue],
-        context: &mut Context,
-    ) -> JsResult<JsValue> {
-        use std::cell::Cell;
-
-        use crate::builtins::iterable::IteratorRecord;
-        use crate::builtins::promise::ResolvingFunctions;
-        use crate::builtins::AsyncFromSyncIterator;
-        use crate::native_function::{ContinuationState, NativeContinuation};
-        use crate::object::{JsFunction, JsPromise};
-        use crate::JsError;
-
-        #[derive(Trace, Finalize)]
-        struct GlobalState {
-            mapfn: Option<JsFunction>,
-            this_arg: JsValue,
-            resolvers: ResolvingFunctions,
-        }
-
-        #[derive(Trace, Finalize)]
-        #[boa_gc(unsafe_no_drop)]
-        enum IterableStateMachine {
-            LoopStart {
-                a: JsObject,
-                k: u64,
-                iterator_record: IteratorRecord,
-            },
-            LoopContinue {
-                a: JsObject,
-                k: u64,
-                iterator_record: IteratorRecord,
-            },
-            LoopEnd {
-                a: JsObject,
-                k: u64,
-                iterator_record: IteratorRecord,
-                mapped_value: Option<JsResult<JsValue>>,
-            },
-            AsyncIteratorCloseStart {
-                err: JsError,
-                iterator: JsObject,
-            },
-            AsyncIteratorCloseEnd {
-                err: JsError,
-            },
-        }
-
-        fn from_iterable(
-            result: JsResult<JsValue>,
-            global_state: &GlobalState,
-            state_machine: &Cell<Option<IterableStateMachine>>,
-            context: &mut Context,
-        ) -> JsResult<ContinuationState> {
-            let Some(mut sm) = state_machine.take() else {
-                return Ok(ContinuationState::Done);
-            };
-
-            loop {
-                match sm {
-                    IterableStateMachine::LoopStart {
-                        a,
-                        k,
-                        iterator_record,
-                    } => {
-                        if k >= 2u64.pow(53) - 1 {
-                            sm = IterableStateMachine::AsyncIteratorCloseStart {
-                                err: JsNativeError::typ()
-                                    .with_message(
-                                        "Array.fromAsync: \
-                                        reached the maximum number of elements in an array \
-                                        (2^53 - 1)",
-                                    )
-                                    .into(),
-                                iterator: iterator_record.iterator().clone(),
-                            };
-                        } else {
-                            let next_result = iterator_record.next_method().call(
-                                &iterator_record.iterator().clone().into(),
-                                &[],
-                                context,
-                            )?;
-
-                            state_machine.set(Some(IterableStateMachine::LoopContinue {
-                                a,
-                                k,
-                                iterator_record,
-                            }));
-                            return Ok(ContinuationState::Yielded(next_result));
-                        }
-                    }
-                    IterableStateMachine::LoopContinue {
-                        a,
-                        k,
-                        mut iterator_record,
-                    } => {
-                        iterator_record.update_result(result.clone()?, context)?;
-
-                        if iterator_record.done() {
-                            a.set(js_string!("length"), k, true, context)?;
-                            global_state
-                                .resolvers
-                                .resolve
-                                .call(&JsValue::undefined(), &[a.into()], context)
-                                .expect("resolving functions cannot fail");
-
-                            return Ok(ContinuationState::Done);
-                        }
-
-                        let next_value = iterator_record.value(context)?;
-                        if let Some(mapfn) = &global_state.mapfn {
-                            let mapped_value = match mapfn.call(
-                                &global_state.this_arg,
-                                &[next_value, k.into()],
-                                context,
-                            ) {
-                                Ok(v) => v,
-                                Err(err) => {
-                                    sm = IterableStateMachine::AsyncIteratorCloseStart {
-                                        err,
-                                        iterator: iterator_record.iterator().clone(),
-                                    };
-                                    continue;
-                                }
-                            };
-                            state_machine.set(Some(IterableStateMachine::LoopEnd {
-                                a,
-                                k,
-                                iterator_record,
-                                mapped_value: None,
-                            }));
-                            return Ok(ContinuationState::Yielded(mapped_value));
-                        }
-                        sm = IterableStateMachine::LoopEnd {
-                            a,
-                            k,
-                            iterator_record,
-                            mapped_value: Some(Ok(next_value)),
-                        }
-                    }
-                    IterableStateMachine::LoopEnd {
-                        a,
-                        k,
-                        iterator_record,
-                        mapped_value,
-                    } => {
-                        let mapped_value = match mapped_value.unwrap_or(result.clone()) {
-                            Ok(value) => value,
-                            Err(err) => {
-                                sm = IterableStateMachine::AsyncIteratorCloseStart {
-                                    err,
-                                    iterator: iterator_record.iterator().clone(),
-                                };
-                                continue;
-                            }
-                        };
-
-                        sm = if let Err(err) =
-                            a.create_data_property_or_throw(k, mapped_value, context)
-                        {
-                            IterableStateMachine::AsyncIteratorCloseStart {
-                                err,
-                                iterator: iterator_record.iterator().clone(),
-                            }
-                        } else {
-                            IterableStateMachine::LoopStart {
-                                a,
-                                k: k + 1,
-                                iterator_record,
-                            }
-                        };
-                    }
-                    IterableStateMachine::AsyncIteratorCloseStart { err, iterator } => {
-                        let Ok(Some(ret)) = iterator.get_method(js_string!("return"), context)
-                        else {
-                            return Err(err);
-                        };
-
-                        let Ok(value) = ret.call(&iterator.into(), &[], context) else {
-                            return Err(err);
-                        };
-
-                        state_machine
-                            .set(Some(IterableStateMachine::AsyncIteratorCloseEnd { err }));
-                        return Ok(ContinuationState::Yielded(value));
-                    }
-                    IterableStateMachine::AsyncIteratorCloseEnd { err } => {
-                        return Err(err);
-                    }
-                }
-            }
-        }
-
-        #[derive(Trace, Finalize)]
-        #[boa_gc(unsafe_no_drop)]
-        enum ArrayLikeStateMachine {
-            LoopStart {
-                array_like: JsObject,
-                a: JsObject,
-                len: u64,
-                k: u64,
-            },
-            LoopContinue {
-                array_like: JsObject,
-                a: JsObject,
-                len: u64,
-                k: u64,
-            },
-            LoopEnd {
-                array_like: JsObject,
-                a: JsObject,
-                len: u64,
-                k: u64,
-                mapped_value: Option<JsValue>,
-            },
-        }
-
-        fn from_array_like(
-            result: JsResult<JsValue>,
-            global_state: &GlobalState,
-            state_machine: &Cell<Option<ArrayLikeStateMachine>>,
-            context: &mut Context,
-        ) -> JsResult<ContinuationState> {
-            let Some(mut sm) = state_machine.take() else {
-                return Ok(ContinuationState::Done);
-            };
-
-            loop {
-                match sm {
-                    ArrayLikeStateMachine::LoopStart {
-                        array_like,
-                        a,
-                        len,
-                        k,
-                    } => {
-                        if k >= len {
-                            a.set(js_string!("length"), len, true, context)?;
-                            global_state
-                                .resolvers
-                                .resolve
-                                .call(&JsValue::undefined(), &[a.into()], context)
-                                .expect("resolving functions cannot fail");
-                            return Ok(ContinuationState::Done);
-                        }
-                        let k_value = array_like.get(k, context)?;
-                        state_machine.set(Some(ArrayLikeStateMachine::LoopContinue {
-                            array_like,
-                            a,
-                            len,
-                            k,
-                        }));
-                        return Ok(ContinuationState::Yielded(k_value));
-                    }
-                    ArrayLikeStateMachine::LoopContinue {
-                        array_like,
-                        a,
-                        len,
-                        k,
-                    } => {
-                        let k_value = result.clone()?;
-                        if let Some(mapfn) = &global_state.mapfn {
-                            let mapped_value = mapfn.call(
-                                &global_state.this_arg,
-                                &[k_value, k.into()],
-                                context,
-                            )?;
-                            state_machine.set(Some(ArrayLikeStateMachine::LoopEnd {
-                                array_like,
-                                a,
-                                len,
-                                k,
-                                mapped_value: None,
-                            }));
-                            return Ok(ContinuationState::Yielded(mapped_value));
-                        }
-                        sm = ArrayLikeStateMachine::LoopEnd {
-                            array_like,
-                            a,
-                            len,
-                            k,
-                            mapped_value: Some(k_value),
-                        }
-                    }
-                    ArrayLikeStateMachine::LoopEnd {
-                        array_like,
-                        a,
-                        len,
-                        k,
-                        mapped_value,
-                    } => {
-                        let mapped_value = mapped_value.unwrap_or(result.clone()?);
-                        a.create_data_property_or_throw(k, mapped_value, context)?;
-                        sm = ArrayLikeStateMachine::LoopStart {
-                            array_like,
-                            a,
-                            len,
-                            k: k + 1,
-                        }
-                    }
-                }
-            }
-        }
-
-        let (promise, resolvers) = JsPromise::new_pending(context);
-
-        let async_items = args.get_or_undefined(0);
-        let mapfn = args.get_or_undefined(1);
-        let this_arg = args.get_or_undefined(2).clone();
-
-        let result: JsResult<()> = (|| {
-            let mapfn = if mapfn.is_undefined() {
-                None
-            } else {
-                let Some(callable) = mapfn.as_callable().cloned() else {
-                    return Err(JsNativeError::typ()
-                        .with_message("Array.fromAsync: mapping function must be callable")
-                        .into());
-                };
-                Some(JsFunction::from_object_unchecked(callable))
-            };
-
-            let iterator_record = if let Some(method) =
-                async_items.get_method(JsSymbol::async_iterator(), context)?
-            {
-                async_items.get_iterator_from_method(&method, context)?
-            } else if let Some(method) = async_items.get_method(JsSymbol::iterator(), context)? {
-                AsyncFromSyncIterator::create(
-                    async_items.get_iterator_from_method(&method, context)?,
-                    context,
-                )
-            } else {
-                let array_like = async_items.to_object(context)?;
-                let len = array_like.length_of_array_like(context)?;
-                let a = if let Some(c) = this.as_constructor() {
-                    c.construct(&[len.into()], None, context)?
-                } else {
-                    Array::array_create(len, None, context)?
-                };
-                let continuation = NativeContinuation::from_copy_closure_with_captures(
-                    |result, (gs, sm), context| match from_array_like(result, gs, sm, context) {
-                        Ok(cont) => cont,
-                        Err(err) => {
-                            gs.resolvers
-                                .reject
-                                .call(&JsValue::undefined(), &[err.to_opaque(context)], context)
-                                .expect("resolving functions cannot fail");
-                            ContinuationState::Done
-                        }
-                    },
-                    (
-                        GlobalState {
-                            mapfn,
-                            this_arg,
-                            resolvers: resolvers.clone(),
-                        },
-                        Cell::new(Some(ArrayLikeStateMachine::LoopStart {
-                            array_like,
-                            a,
-                            len,
-                            k: 0,
-                        })),
-                    ),
-                );
-
-                if let ContinuationState::Yielded(value) =
-                    continuation.call(Ok(JsValue::undefined()), context)
-                {
-                    JsPromise::resolve(value, context).await_native(continuation, context);
-                }
-
-                return Ok(());
-            };
-
-            let a = if let Some(c) = this.as_constructor() {
-                c.construct(&[], None, context)?
-            } else {
-                Array::array_create(0, None, context)?
-            };
-
-            let continuation = NativeContinuation::from_copy_closure_with_captures(
-                |result, (gs, sm), context| match from_iterable(result, gs, sm, context) {
-                    Ok(cont) => cont,
-                    Err(err) => {
-                        gs.resolvers
-                            .reject
-                            .call(&JsValue::undefined(), &[err.to_opaque(context)], context)
-                            .expect("resolving functions cannot fail");
-                        ContinuationState::Done
-                    }
-                },
-                (
-                    GlobalState {
-                        mapfn,
-                        this_arg,
-                        resolvers: resolvers.clone(),
-                    },
-                    Cell::new(Some(IterableStateMachine::LoopStart {
-                        a,
-                        k: 0,
-                        iterator_record,
-                    })),
-                ),
-            );
-
-            if let ContinuationState::Yielded(value) =
-                continuation.call(Ok(JsValue::undefined()), context)
-            {
-                JsPromise::resolve(value, context).await_native(continuation, context);
-            }
-
-            Ok(())
-        })();
-
-        if let Err(err) = result {
-            resolvers
-                .reject
-                .call(&JsValue::undefined(), &[err.to_opaque(context)], context)
-                .expect("resolving functions cannot fail");
-        }
-
-        return Ok(promise.into());
     }
 
     /// `Array.isArray( arg )`
