+++ core/engine/src/native_function/continuation.rs
@@ -4,22 +4,22 @@ use crate::{Context, JsResult, JsValue};
 
 #[derive(Trace, Finalize)]
 #[boa_gc(unsafe_no_drop)]
-pub(crate) enum ContinuationState {
+pub(crate) enum CoroutineState {
     Yielded(JsValue),
     Done,
 }
 
-trait TraceableContinuation: Trace {
-    fn call(&self, value: JsResult<JsValue>, context: &mut Context) -> ContinuationState;
+trait TraceableCoroutine: Trace {
+    fn call(&self, value: JsResult<JsValue>, context: &mut Context) -> CoroutineState;
 }
 
 #[derive(Trace, Finalize)]
-struct Continuation<F, T>
+struct Coroutine<F, T>
 where
-    F: Fn(JsResult<JsValue>, &T, &mut Context) -> ContinuationState,
+    F: Fn(JsResult<JsValue>, &T, &mut Context) -> CoroutineState,
     T: Trace,
 {
-    // SAFETY: `NativeContinuation`'s safe API ensures only `Copy` closures are stored; its unsafe API,
+    // SAFETY: `NativeCoroutine`'s safe API ensures only `Copy` closures are stored; its unsafe API,
     // on the other hand, explains the invariants to hold in order for this to be safe, shifting
     // the responsibility to the caller.
     #[unsafe_ignore_trace]
@@ -27,48 +27,48 @@ where
     captures: T,
 }
 
-impl<F, T> TraceableContinuation for Continuation<F, T>
+impl<F, T> TraceableCoroutine for Coroutine<F, T>
 where
-    F: Fn(JsResult<JsValue>, &T, &mut Context) -> ContinuationState,
+    F: Fn(JsResult<JsValue>, &T, &mut Context) -> CoroutineState,
     T: Trace,
 {
-    fn call(&self, result: JsResult<JsValue>, context: &mut Context) -> ContinuationState {
+    fn call(&self, result: JsResult<JsValue>, context: &mut Context) -> CoroutineState {
         (self.f)(result, &self.captures, context)
     }
 }
 
-/// A callable Rust continuation that can be used to await promises.
+/// A callable Rust coroutine that can be used to await promises.
 ///
 /// # Caveats
 ///
 /// By limitations of the Rust language, the garbage collector currently cannot inspect closures
 /// in order to trace their captured variables. This means that only [`Copy`] closures are 100% safe
-/// to use. All other closures can also be stored in a `NativeContinuation`, albeit by using an `unsafe`
+/// to use. All other closures can also be stored in a `NativeCoroutine`, albeit by using an `unsafe`
 /// API, but note that passing closures implicitly capturing traceable types could cause
 /// **Undefined Behaviour**.
 #[derive(Clone, Trace, Finalize)]
-pub(crate) struct NativeContinuation {
-    inner: Gc<dyn TraceableContinuation>,
+pub(crate) struct NativeCoroutine {
+    inner: Gc<dyn TraceableCoroutine>,
 }
 
-impl std::fmt::Debug for NativeContinuation {
+impl std::fmt::Debug for NativeCoroutine {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("NativeContinuation").finish_non_exhaustive()
+        f.debug_struct("NativeCoroutine").finish_non_exhaustive()
     }
 }
 
-impl NativeContinuation {
-    /// Creates a `NativeFunction` from a `Copy` closure and a list of traceable captures.
+impl NativeCoroutine {
+    /// Creates a `NativeCoroutine` from a `Copy` closure and a list of traceable captures.
     pub(crate) fn from_copy_closure_with_captures<F, T>(closure: F, captures: T) -> Self
     where
-        F: Fn(JsResult<JsValue>, &T, &mut Context) -> ContinuationState + Copy + 'static,
+        F: Fn(JsResult<JsValue>, &T, &mut Context) -> CoroutineState + Copy + 'static,
         T: Trace + 'static,
     {
         // SAFETY: The `Copy` bound ensures there are no traceable types inside the closure.
         unsafe { Self::from_closure_with_captures(closure, captures) }
     }
 
-    /// Create a new `NativeFunction` from a closure and a list of traceable captures.
+    /// Create a new `NativeCoroutine` from a closure and a list of traceable captures.
     ///
     /// # Safety
     ///
@@ -78,12 +78,12 @@ impl NativeContinuation {
     /// on why that is the case.
     pub(crate) unsafe fn from_closure_with_captures<F, T>(closure: F, captures: T) -> Self
     where
-        F: Fn(JsResult<JsValue>, &T, &mut Context) -> ContinuationState + 'static,
+        F: Fn(JsResult<JsValue>, &T, &mut Context) -> CoroutineState + 'static,
         T: Trace + 'static,
     {
         // Hopefully, this unsafe operation will be replaced by the `CoerceUnsized` API in the
         // future: https://github.com/rust-lang/rust/issues/18598
-        let ptr = Gc::into_raw(Gc::new(Continuation {
+        let ptr = Gc::into_raw(Gc::new(Coroutine {
             f: closure,
             captures,
         }));
@@ -96,13 +96,9 @@ impl NativeContinuation {
         }
     }
 
-    /// Calls this `NativeFunction`, forwarding the arguments to the corresponding function.
+    /// Calls this `NativeCoroutine`, forwarding the arguments to the corresponding function.
     #[inline]
-    pub(crate) fn call(
-        &self,
-        result: JsResult<JsValue>,
-        context: &mut Context,
-    ) -> ContinuationState {
+    pub(crate) fn call(&self, result: JsResult<JsValue>, context: &mut Context) -> CoroutineState {
         self.inner.call(result, context)
     }
 }
